+++ applications/services/gui/icon_i.h
@@ -7,8 +7,8 @@
 #include "icon.h"
 
 struct Icon {
-    const uint8_t width;
-    const uint8_t height;
+    const uint16_t width;
+    const uint16_t height;
     const uint8_t frame_count;
     const uint8_t frame_rate;
     const uint8_t* const* frames;
